requirements.txt
# FastAPI 相關
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# 資料庫
pymysql==1.1.0
sqlalchemy==2.0.23
alembic==1.12.1

# 認證
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI/影像處理 (更新版本以支援 Python 3.12)
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.26.0
torch>=2.1.0
torchvision>=0.16.0

# 其他工具
Pillow>=10.0.0
aiofiles==23.2.1